refactor: enhance file location handling
Introduce `FileLocationInitializer` to manage file location initialization, improving modularity and code organization. Update function signatures and imports to support the new initializer.
This refactoring aims to enhance maintainability and scalability by decoupling file location logic from the core module.
