+++ src/libs/feature/feature-file-location-initializer/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/file-location-initializer';
